+++ tests/list.c
@@ -20,7 +20,7 @@
 #include <yatl/lite.h>
 #include <libmarias3/marias3.h>
 
-/* Tests basic put, list, get, status, delete using the thread calls */
+/* Tests list command */
 
 int main(int argc, char *argv[])
 {
@@ -28,10 +28,7 @@ int main(int argc, char *argv[])
   (void) argv;
   int res;
   ms3_list_st *list = NULL, *list_it = NULL;
-  uint8_t *data;
-  size_t length;
   const char *test_string = "Another one bites the dust";
-  ms3_status_st status;
   char *s3key = getenv("S3KEY");
   char *s3secret = getenv("S3SECRET");
   char *s3region = getenv("S3REGION");
@@ -56,40 +53,45 @@ int main(int argc, char *argv[])
 //  ms3_debug(true);
   ASSERT_NOT_NULL(ms3);
 
-  res = ms3_put(ms3, s3bucket, "test/basic_thread.txt",
+  res = ms3_put(ms3, s3bucket, "list1/test1.txt",
+                (const uint8_t *)test_string,
+                strlen(test_string));
+  ASSERT_EQ_(res, 0, "Result: %u", res);
+
+  res = ms3_put(ms3, s3bucket, "list2/test2.txt",
                 (const uint8_t *)test_string,
                 strlen(test_string));
   ASSERT_EQ_(res, 0, "Result: %u", res);
 
   // A prefix that will give no results;
-  res = ms3_list(ms3, s3bucket, "asdfghjkl", &list);
+  res = ms3_list_dir(ms3, s3bucket, "asdfghjkl", &list);
   ASSERT_EQ_(res, 0, "Result: %u", res);
   ASSERT_NULL_(list, "List not empty");
 
-  res = ms3_list(ms3, s3bucket, NULL, &list);
+  res = ms3_list_dir(ms3, s3bucket, NULL, &list);
   ASSERT_EQ_(res, 0, "Result: %u", res);
   bool found = false;
+  bool found_bad = false;
   list_it = list;
 
   while (list_it)
   {
-    if (!strncmp(list_it->key, "test/basic_thread.txt", 21))
+    if (!strncmp(list_it->key, "list1/test1.txt", 15))
+    {
+      found_bad = true;
+    }
+    if (!strncmp(list_it->key, "list2/", 6))
     {
       found = true;
-      break;
     }
 
     list_it = list_it->next;
   }
 
   ASSERT_EQ_(found, 1, "Created file not found");
+  ASSERT_NEQ_(found_bad, 1, "File listed should not be here");
 
-  if (list_it)
-  {
-    ASSERT_EQ_(list_it->length, 26, "Created file is unexpected length");
-    ASSERT_NEQ_(list_it->created, 0, "Created file timestamp is bad");
-  }
-  else
+  if (!list)
   {
     ASSERT_TRUE_(false, "No resuts from list");
   }
@@ -100,67 +102,40 @@ int main(int argc, char *argv[])
   uint8_t list_version = 1;
   list = NULL;
   ms3_set_option(ms3, MS3_OPT_FORCE_LIST_VERSION, &list_version);
-  res = ms3_list(ms3, s3bucket, NULL, &list);
+  res = ms3_list_dir(ms3, s3bucket, NULL, &list);
   ASSERT_EQ_(res, 0, "Result: %u", res);
   found = false;
   list_it = list;
 
   while (list_it)
   {
-    if (!strncmp(list_it->key, "test/basic_thread.txt", 21))
+    if (!strncmp(list_it->key, "list1/test1.txt", 15))
+    {
+      found_bad = true;
+    }
+    if (!strncmp(list_it->key, "list2/", 6))
     {
       found = true;
-      break;
     }
 
     list_it = list_it->next;
   }
 
   ASSERT_EQ_(found, 1, "Created file not found");
+  ASSERT_NEQ_(found_bad, 1, "File listed should not be here");
 
-  if (list_it)
-  {
-    ASSERT_EQ_(list_it->length, 26, "Created file is unexpected length");
-    ASSERT_NEQ_(list_it->created, 0, "Created file timestamp is bad");
-  }
-  else
+  if (!list)
   {
     ASSERT_TRUE_(false, "No resuts from list");
   }
 
   ms3_list_free(list);
 
-  res = ms3_get(ms3, s3bucket, "test/basic_thread.txt", &data, &length);
+  res = ms3_delete(ms3, s3bucket, "test1/test1.txt");
   ASSERT_EQ_(res, 0, "Result: %u", res);
-  ASSERT_EQ(length, 26);
-  ASSERT_STREQ((char *)data, test_string);
-
-  for (int i = 0; i <= 3; i++)
-  {
-    res = ms3_status(ms3, s3bucket, "test/basic_thread.txt", &status);
-
-    if (res == MS3_ERR_NOT_FOUND)
-    {
-      continue;
-    }
-
-    ASSERT_EQ_(res, 0, "Result: %u", res);
-
-    if (res == 0)
-    {
-      break;
-    }
-  }
-
-  ASSERT_EQ(status.length, 26);
-  ASSERT_NEQ(status.created, 0);
-  res = ms3_delete(ms3, s3bucket, "test/basic_thread.txt");
+  res = ms3_delete(ms3, s3bucket, "test2/test2.txt");
   ASSERT_EQ_(res, 0, "Result: %u", res);
-  ms3_free(data);
-  res = ms3_get(ms3, s3bucket, "test/basic_thread.txt", &data, &length);
-  ASSERT_NEQ_(res, 0, "Object should error");
-  ASSERT_NULL_(data, "Data should be NULL");
-  ASSERT_EQ_(length, 0, "There should be no data");
+
   ms3_deinit(ms3);
   ms3_library_deinit();
 }